backend/requirements-complete.txt
# AURA OSINT Complete Dependencies - All Tools Included

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Vector Database
qdrant-client==1.7.0
sentence-transformers==2.2.2

# OSINT Tools - Email
holehe==1.65
h8mail==2.5.6
emailrep==0.1.3

# OSINT Tools - Social Media
ntscraper==0.4.0
instaloader==4.10
tweepy==4.14.0
praw==7.7.1
telethon==1.34.0

# OSINT Tools - Network
shodan==1.31.0
python-nmap==0.7.1
dnspython==2.4.2
python-whois==0.8.0
censys==2.2.8

# OSINT Tools - Phone
phonenumbers==8.13.26
phonenumberslite==8.13.26

# OSINT Tools - Darknet/Tor
stem==1.8.2
PySocks==1.7.1
requests[socks]==2.31.0
torrequest==0.1.0

# OSINT Tools - Crypto/Blockchain
blockcypher==1.0.93
web3==6.11.3
bitcoin==1.1.42
monero==1.1.1
blockchain==1.4.4

# OSINT Tools - Image/Face Recognition
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.1.0
imagehash==4.3.1

# OSINT Tools - General
googlesearch-python==1.2.3
duckduckgo-search==3.9.6
waybackpy==3.0.6
sublist3r==1.1
theHarvester==4.5.1

# Machine Learning & AI
scikit-learn==1.3.2
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.36.0
torch==2.1.1
numpy==1.24.4
pandas==2.1.4

# Web Scraping & HTTP
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
playwright==1.40.0

# Security & Crypto
cryptography==41.0.8
passlib==4.1.2
bcrypt==4.1.2
python-jose==3.3.0
pyotp==2.9.0

# Monitoring & Logging
psutil==5.9.6
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0
user-agents==2.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: GPU acceleration (uncomment if needed)
# torch-audio==2.1.1
# torchvision==0.16.1